chore(item): extract helper functions into a service
